requirements.txt
torch
flask
pillow
sentence-transformers
matplotlib
numpy
scipy